refactor(examples): 迁移 stencil 至 plugins 目录
