+++ examples/src/App.tsx
@@ -38,7 +38,7 @@ import { KeyboardExample } from './pages/plugins/keyboard'
 import { DndExample } from './pages/plugins/dnd'
 import { ScrollerExample } from './pages/plugins/scroller'
 import { SelectionExample } from './pages/plugins/selection'
-import { StencilExample } from './pages/stencil'
+import { StencilExample } from './pages/plugins/stencil'
 import { TransformExample } from './pages/transform'
 import { UndoExample } from './pages/undo'
 import { TransitionExample } from './pages/animation/transition'
@@ -96,8 +96,8 @@ function App() {
       <Route path="/plugins/keyboard" element={<KeyboardExample />} />
       <Route path="/plugins/scroller" element={<ScrollerExample />} />
       <Route path="/plugins/selection" element={<SelectionExample />} />
+      <Route path="/plugins/stencil" element={<StencilExample />} />
 
-      <Route path="/stencil" element={<StencilExample />} />
       <Route path="/transform" element={<TransformExample />} />
       <Route path="/undo" element={<UndoExample />} />
       <Route path="/animation/transition" element={<TransitionExample />} />
