+++ examples/src/pages/plugins/stencil/index.less
